# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManagemenLogbook API",
        "description": "API documentation for ManagemenLogbook",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tugas": {
            "get": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Get tasks (filter & search)",
                "operationId": "cd8850d85fa266867855fbba951bdccf",
                "parameters": [
                    {
                        "name": "kgt_id",
                        "in": "query",
                        "description": "Filter Kegiatan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "usr_id",
                        "in": "query",
                        "description": "Filter PIC",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter Status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search Name/Code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Create a tugas with auto-generated prefix",
                "operationId": "59ee41cbceba1efec3361389544684d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kgt_id",
                                    "usr_id",
                                    "nama",
                                    "tgl_mulai",
                                    "tgl_selesai",
                                    "bobot"
                                ],
                                "properties": {
                                    "kgt_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usr_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Integrasi API Payment"
                                    },
                                    "tgl_mulai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-15"
                                    },
                                    "tgl_selesai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-20"
                                    },
                                    "bobot": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tugas created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tugas/{id}": {
            "get": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Get tugas detail",
                "operationId": "32fdd439e52aa6a4a9aaf3273d64502e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tugas details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Update tugas (Progress or Full Data)",
                "operationId": "b79338bc76e9a4a2bd346e35e6449a3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "progress": {
                                        "type": "number"
                                    },
                                    "nama": {
                                        "type": "string"
                                    },
                                    "usr_id": {
                                        "type": "integer"
                                    },
                                    "tgl_selesai": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tugas updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Delete tugas",
                "operationId": "832edc4587778f7e4765aa17cf0b012f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tugas deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}